Adding a check for domain.

+++ cfme/tests/infrastructure/test_quota_tagging.py
@@ -52,7 +52,8 @@ def test_domain(appliance):
                                                       fauxfactory.gen_alphanumeric()),
                                                   enabled=True)
     yield domain
-    domain.delete()
+    if domain.exists:
+        domain.delete()
 
 
 @pytest.fixture(scope='module')
